Created SlaEndpoint class and refactored SLA class
