+++ src/main/java/agh/edu/zeuspol/datastructures/storages/Sla.java
@@ -1,21 +1,25 @@
-package agh.edu.zeuspol.datastructures;
+package agh.edu.zeuspol.datastructures.storages;
+
+import agh.edu.zeuspol.datastructures.NotificationRule;
+import agh.edu.zeuspol.datastructures.Rule;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class Sla {
 
+    private static final Sla instance = new Sla();
+
     private final List<Rule> rules;
     private final List<NotificationRule> notificationRules;
 
-    public Sla(){
+    private Sla(){
         this.rules = new ArrayList<>();
         this.notificationRules = new ArrayList<>();
     }
 
-    public Sla(List<Rule> rules, List<NotificationRule> notificationRules){
-        this.rules = rules;
-        this.notificationRules = notificationRules;
+    public static Sla getInstance(){
+        return instance;
     }
 
     public void addRules(List<Rule> rules){
